requirements.txt
allure-pytest==2.13.2
allure-python-common==2.13.2
attrs==23.1.0
certifi==2023.7.22
execnet==2.0.2
h11==0.14.0
inda==3.4
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
PySocks==1.7.1
pytest==7.4.3
pytest-xdist==3.3.1
selenium==4.15.2
sniffio==1.3.0
sortedcontaines==2.4.0
trio-websocket==0.11.1
urllib==2.0.7
wsproto==1.2.0